requirements.txt
flask>=0.10.1
flask-login
flask-openid
flask-mail
flask-sqlalchemy
sqlalchemy-migrate
pymysql
flask-whooshalchemy
flask-wtf
flask-babel
guess_language
flipflop
coverage
Flask-Migrate
PyJWT
cryptography
flask-cors
click
